feat: 관리자의 예약 대기 관리 추가

+++ src/main/java/roomescape/dto/ReservationWaitingResponse.java
@@ -3,5 +3,5 @@ package roomescape.dto;
 import java.time.LocalDate;
 
 public record ReservationWaitingResponse(long id, String name, LocalDate date, ReservationTimeResponse time,
-                                         ThemeResponse theme, long priority) {
+                                         ThemeResponse theme, Integer priority) {
 }
